+++ app/ui/layout/header.tsx
@@ -1,4 +1,4 @@
-import Logo from "./logo";
+import Logo from "../logo";
 
 export default function Header() {
   return (
