+++ app/ui/layout/internal-navbar.tsx
@@ -1,4 +1,4 @@
-import { NavLink } from "../lib/definitions";
+import { NavLink } from "../../lib/definitions";
 import Link from "next/link";
 import { usePathname } from "next/navigation";
